+++ Cargo.toml
@@ -3,6 +3,7 @@ members = [
   "creusot",
   "creusot-contracts",
   "creusot-contracts-proc",
+  "creusot-contracts-dummy",
   "creusot-metadata",
   "why3",
   "why3tests",
