+++ creusot-contracts-dummy/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "creusot-contracts-dummy"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+proc-macro = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
