+++ creusot-contracts/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2018"
 
 [dependencies]
 creusot-contracts-proc = { path = "../creusot-contracts-proc", version = "*" }
+creusot-contracts-dummy = { path = "../creusot-contracts-dummy", version = "*" }
 rand = "*"
 
 [features]
