+++ creusot-contracts/src/logic.rs
@@ -1,4 +1,5 @@
 pub mod eq;
+mod ghost;
 mod int;
 mod model;
 pub mod ord;
@@ -7,6 +8,7 @@ mod seq;
 pub mod well_founded;
 
 pub use eq::*;
+pub use ghost::*;
 pub use int::*;
 pub use model::*;
 pub use ord::*;
