+++ creusot/tests/should_succeed/iter_mut.stdout
@@ -45,8 +45,8 @@ module Type
     ensures { result = core_option_option_Some_0 self }
     
   axiom core_option_option_Some_0_acc : forall a : 't . core_option_option_Some_0 (Core_Option_Option_Some a : core_option_option 't) = a
-  type itermut_ghost 't = 
-    | IterMut_Ghost opaque_ptr
+  type creusotcontracts_logic_ghost_ghost 't = 
+    | CreusotContracts_Logic_Ghost_Ghost opaque_ptr
     
 end
 module CreusotContracts_Logic_Model_Model_ModelTy
@@ -226,49 +226,6 @@ module IterMut_Impl3_Next
     ensures { result = Get0.get (Model0.model ( * self)) 0 }
     
 end
-module IterMut_Impl4_ModelTy
-  type t   
-  type modelTy  = 
-    t
-end
-module IterMut_Impl4_Model_Interface
-  type t   
-  use Type
-  function model (self : Type.itermut_ghost t) : t
-end
-module IterMut_Impl4_Model
-  type t   
-  use Type
-  function model (self : Type.itermut_ghost t) : t
-end
-module IterMut_Impl4
-  type t   
-  use Type
-  clone IterMut_Impl4_Model as Model0 with type t = t
-  clone IterMut_Impl4_ModelTy as ModelTy0 with type t = t
-  clone CreusotContracts_Logic_Model_Model_Model as Model1 with type self = Type.itermut_ghost t,
-  type ModelTy0.modelTy = ModelTy0.modelTy, function model = Model0.model
-  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy1 with type self = Type.itermut_ghost t,
-  type modelTy = ModelTy0.modelTy
-end
-module IterMut_Impl5_Record_Interface
-  type t   
-  use prelude.Prelude
-  use Type
-  clone IterMut_Impl4_Model_Interface as Model0 with type t = t
-  val record (a : t) : Type.itermut_ghost t
-    ensures { Model0.model result = a }
-    
-end
-module IterMut_Impl5_Record
-  type t   
-  use prelude.Prelude
-  use Type
-  clone IterMut_Impl4_Model as Model0 with type t = t
-  val record (a : t) : Type.itermut_ghost t
-    ensures { Model0.model result = a }
-    
-end
 module CreusotContracts_Logic_Model_Impl1_ModelTy
   type t   
   clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = t
@@ -303,6 +260,31 @@ module CreusotContracts_Logic_Model_Impl1
   clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy1 with type self = borrowed t,
   type modelTy = ModelTy0.modelTy
 end
+module CreusotContracts_Logic_Ghost_Impl0_ModelTy
+  type t   
+  type modelTy  = 
+    t
+end
+module CreusotContracts_Logic_Ghost_Impl0_Model_Interface
+  type t   
+  use Type
+  function model (self : Type.creusotcontracts_logic_ghost_ghost t) : t
+end
+module CreusotContracts_Logic_Ghost_Impl0_Model
+  type t   
+  use Type
+  function model (self : Type.creusotcontracts_logic_ghost_ghost t) : t
+end
+module CreusotContracts_Logic_Ghost_Impl0
+  type t   
+  use Type
+  clone CreusotContracts_Logic_Ghost_Impl0_Model as Model0 with type t = t
+  clone CreusotContracts_Logic_Ghost_Impl0_ModelTy as ModelTy0 with type t = t
+  clone CreusotContracts_Logic_Model_Model_Model as Model1 with type self = Type.creusotcontracts_logic_ghost_ghost t,
+  type ModelTy0.modelTy = ModelTy0.modelTy, function model = Model0.model
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy1 with type self = Type.creusotcontracts_logic_ghost_ghost t,
+  type modelTy = ModelTy0.modelTy
+end
 module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
@@ -310,6 +292,24 @@ end
 module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
+end
+module CreusotContracts_Logic_Ghost_Impl1_Record_Interface
+  type t   
+  use prelude.Prelude
+  use Type
+  clone CreusotContracts_Logic_Ghost_Impl0_Model_Interface as Model0 with type t = t
+  val record (a : t) : Type.creusotcontracts_logic_ghost_ghost t
+    ensures { Model0.model result = a }
+    
+end
+module CreusotContracts_Logic_Ghost_Impl1_Record
+  type t   
+  use prelude.Prelude
+  use Type
+  clone CreusotContracts_Logic_Ghost_Impl0_Model as Model0 with type t = t
+  val record (a : t) : Type.creusotcontracts_logic_ghost_ghost t
+    ensures { Model0.model result = a }
+    
 end
 module CreusotContracts_Logic_Resolve_Impl1_Resolve_Interface
   type t   
@@ -353,7 +353,7 @@ module IterMut_IncVec
   clone IterMut_Impl0_Model as Model1 with type t = int
   clone CreusotContracts_Logic_Model_Impl1_Model as Model2 with type t = Type.itermut_vec int,
   type ModelTy0.modelTy = ModelTy0.modelTy, function Model0.model = Model1.model
-  clone IterMut_Impl4_Model as Model0 with type t = borrowed (Type.itermut_vec int)
+  clone CreusotContracts_Logic_Ghost_Impl0_Model as Model0 with type t = borrowed (Type.itermut_vec int)
   clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve7 with type self = ()
   clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve6 with type t = int
   clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve5 with type self = Type.core_option_option (borrowed int)
@@ -362,13 +362,13 @@ module IterMut_IncVec
   clone CreusotContracts_Logic_Seq_Impl0_Get as Get0 with type t = borrowed int
   clone CreusotContracts_Logic_Seq_Impl0_Tail as Tail0 with type t = borrowed int
   clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve2 with type t = Type.itermut_vec int
-  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = Type.itermut_ghost (borrowed (Type.itermut_vec int))
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = Type.creusotcontracts_logic_ghost_ghost (borrowed (Type.itermut_vec int))
   clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = borrowed (Type.itermut_vec int)
   clone IterMut_Impl3_Next_Interface as Next0 with type t = int, function Model0.model = Model3.model,
   function Tail0.tail = Tail0.tail, function Get0.get = Get0.get
   clone IterMut_Impl1_IterMut_Interface as IterMut0 with type t = int, function Model0.model = Model1.model,
   function Model1.model = Model3.model
-  clone IterMut_Impl5_Record_Interface as Record0 with type t = borrowed (Type.itermut_vec int),
+  clone CreusotContracts_Logic_Ghost_Impl1_Record_Interface as Record0 with type t = borrowed (Type.itermut_vec int),
   function Model0.model = Model0.model
   let rec cfg inc_vec (v : borrowed (Type.itermut_vec int)) : ()
     ensures { forall i : (int) . 0 <= i && i < Seq.length (Model1.model ( ^ v)) -> Seq.get (Model1.model ( ^ v)) i = Seq.get (Model2.model v) i + 5 }
@@ -377,7 +377,7 @@ module IterMut_IncVec
    = 
   var _0 : ();
   var v_1 : borrowed (Type.itermut_vec int);
-  var old_v_2 : Type.itermut_ghost (borrowed (Type.itermut_vec int));
+  var old_v_2 : Type.creusotcontracts_logic_ghost_ghost (borrowed (Type.itermut_vec int));
   var _3 : borrowed (Type.itermut_vec int);
   var _4 : borrowed (Type.itermut_vec int);
   var it_5 : Type.itermut_itermut int;
