+++ creusot/tests/should_succeed/vector/01.stdout
@@ -9,9 +9,6 @@ module Type
   use floating_point.Single
   use floating_point.Double
   use prelude.Prelude
-  type c01_ghost 't = 
-    | C01_Ghost
-    
   type core_marker_phantomdata 't = 
     | Core_Marker_PhantomData
     
@@ -29,6 +26,9 @@ module Type
     
   type creusotcontracts_std1_vec_vec 't = 
     | CreusotContracts_Std1_Vec_Vec (alloc_vec_vec 't (alloc_alloc_global))
+    
+  type creusotcontracts_logic_ghost_ghost 't = 
+    | CreusotContracts_Logic_Ghost_Ghost opaque_ptr
     
 end
 module C01_Main_Interface
@@ -60,49 +60,6 @@ module CreusotContracts_Logic_Model_Model_Model
   clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = self
   function model (self : self) : ModelTy0.modelTy
 end
-module C01_Impl0_ModelTy
-  type t   
-  type modelTy  = 
-    t
-end
-module C01_Impl0_Model_Interface
-  type t   
-  use Type
-  function model (self : Type.c01_ghost t) : t
-end
-module C01_Impl0_Model
-  type t   
-  use Type
-  function model (self : Type.c01_ghost t) : t
-end
-module C01_Impl0
-  type t   
-  use Type
-  clone C01_Impl0_Model as Model0 with type t = t
-  clone C01_Impl0_ModelTy as ModelTy0 with type t = t
-  clone CreusotContracts_Logic_Model_Model_Model as Model1 with type self = Type.c01_ghost t,
-  type ModelTy0.modelTy = ModelTy0.modelTy, function model = Model0.model
-  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy1 with type self = Type.c01_ghost t,
-  type modelTy = ModelTy0.modelTy
-end
-module C01_Impl1_Record_Interface
-  type t   
-  use prelude.Prelude
-  use Type
-  clone C01_Impl0_Model_Interface as Model0 with type t = t
-  val record (a : t) : Type.c01_ghost t
-    ensures { Model0.model result = a }
-    
-end
-module C01_Impl1_Record
-  type t   
-  use prelude.Prelude
-  use Type
-  clone C01_Impl0_Model as Model0 with type t = t
-  val record (a : t) : Type.c01_ghost t
-    ensures { Model0.model result = a }
-    
-end
 module CreusotContracts_Std1_Vec_Impl0_ModelTy
   type t   
   use seq.Seq
@@ -131,6 +88,31 @@ module CreusotContracts_Std1_Vec_Impl0
   clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy1 with type self = Type.creusotcontracts_std1_vec_vec t,
   type modelTy = ModelTy0.modelTy
 end
+module CreusotContracts_Logic_Ghost_Impl0_ModelTy
+  type t   
+  type modelTy  = 
+    t
+end
+module CreusotContracts_Logic_Ghost_Impl0_Model_Interface
+  type t   
+  use Type
+  function model (self : Type.creusotcontracts_logic_ghost_ghost t) : t
+end
+module CreusotContracts_Logic_Ghost_Impl0_Model
+  type t   
+  use Type
+  function model (self : Type.creusotcontracts_logic_ghost_ghost t) : t
+end
+module CreusotContracts_Logic_Ghost_Impl0
+  type t   
+  use Type
+  clone CreusotContracts_Logic_Ghost_Impl0_Model as Model0 with type t = t
+  clone CreusotContracts_Logic_Ghost_Impl0_ModelTy as ModelTy0 with type t = t
+  clone CreusotContracts_Logic_Model_Model_Model as Model1 with type self = Type.creusotcontracts_logic_ghost_ghost t,
+  type ModelTy0.modelTy = ModelTy0.modelTy, function model = Model0.model
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy1 with type self = Type.creusotcontracts_logic_ghost_ghost t,
+  type modelTy = ModelTy0.modelTy
+end
 module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
@@ -138,6 +120,24 @@ end
 module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
+end
+module CreusotContracts_Logic_Ghost_Impl1_Record_Interface
+  type t   
+  use prelude.Prelude
+  use Type
+  clone CreusotContracts_Logic_Ghost_Impl0_Model_Interface as Model0 with type t = t
+  val record (a : t) : Type.creusotcontracts_logic_ghost_ghost t
+    ensures { Model0.model result = a }
+    
+end
+module CreusotContracts_Logic_Ghost_Impl1_Record
+  type t   
+  use prelude.Prelude
+  use Type
+  clone CreusotContracts_Logic_Ghost_Impl0_Model as Model0 with type t = t
+  val record (a : t) : Type.creusotcontracts_logic_ghost_ghost t
+    ensures { Model0.model result = a }
+    
 end
 module CreusotContracts_Logic_Model_Impl0_ModelTy
   type t   
@@ -377,13 +377,13 @@ module C01_AllZero
   use prelude.Prelude
   use Type
   clone CreusotContracts_Std1_Vec_Impl0_Model as Model1 with type t = uint32
-  clone C01_Impl0_Model as Model0 with type t = borrowed (Type.creusotcontracts_std1_vec_vec uint32)
+  clone CreusotContracts_Logic_Ghost_Impl0_Model as Model0 with type t = borrowed (Type.creusotcontracts_std1_vec_vec uint32)
   clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve5 with type t = Type.creusotcontracts_std1_vec_vec uint32
   clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve4 with type self = ()
   clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve3 with type t = uint32
   clone CreusotContracts_Std1_Vec_Impl0_ModelTy as ModelTy0 with type t = uint32
   clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve2 with type self = usize
-  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = Type.c01_ghost (borrowed (Type.creusotcontracts_std1_vec_vec uint32))
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = Type.creusotcontracts_logic_ghost_ghost (borrowed (Type.creusotcontracts_std1_vec_vec uint32))
   clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = borrowed (Type.creusotcontracts_std1_vec_vec uint32)
   clone CreusotContracts_Logic_Model_Impl1_Model as Model3 with type t = Type.creusotcontracts_std1_vec_vec uint32,
   type ModelTy0.modelTy = ModelTy0.modelTy, function Model0.model = Model1.model
@@ -392,7 +392,7 @@ module C01_AllZero
   clone CreusotContracts_Logic_Model_Impl0_Model as Model2 with type t = Type.creusotcontracts_std1_vec_vec uint32,
   type ModelTy0.modelTy = ModelTy0.modelTy, function Model0.model = Model1.model
   clone CreusotContracts_Std1_Vec_Impl1_Len_Interface as Len0 with type t = uint32, function Model0.model = Model2.model
-  clone C01_Impl1_Record_Interface as Record0 with type t = borrowed (Type.creusotcontracts_std1_vec_vec uint32),
+  clone CreusotContracts_Logic_Ghost_Impl1_Record_Interface as Record0 with type t = borrowed (Type.creusotcontracts_std1_vec_vec uint32),
   function Model0.model = Model0.model
   let rec cfg all_zero (v : borrowed (Type.creusotcontracts_std1_vec_vec uint32)) : ()
     ensures { Seq.length (Model1.model ( * v)) = Seq.length (Model1.model ( ^ v)) }
@@ -402,7 +402,7 @@ module C01_AllZero
   var _0 : ();
   var v_1 : borrowed (Type.creusotcontracts_std1_vec_vec uint32);
   var i_2 : usize;
-  var old_v_3 : Type.c01_ghost (borrowed (Type.creusotcontracts_std1_vec_vec uint32));
+  var old_v_3 : Type.creusotcontracts_logic_ghost_ghost (borrowed (Type.creusotcontracts_std1_vec_vec uint32));
   var _4 : borrowed (Type.creusotcontracts_std1_vec_vec uint32);
   var _5 : borrowed (Type.creusotcontracts_std1_vec_vec uint32);
   var _6 : ();
