+++ creusot/tests/should_succeed/vector/02_gnome.stdout
@@ -9,9 +9,6 @@ module Type
   use floating_point.Single
   use floating_point.Double
   use prelude.Prelude
-  type c02gnome_ghost 't = 
-    | C02Gnome_Ghost opaque_ptr
-    
   type core_cmp_ordering  = 
     | Core_Cmp_Ordering_Less
     | Core_Cmp_Ordering_Equal
@@ -35,62 +32,8 @@ module Type
   type creusotcontracts_std1_vec_vec 't = 
     | CreusotContracts_Std1_Vec_Vec (alloc_vec_vec 't (alloc_alloc_global))
     
-end
-module CreusotContracts_Logic_Model_Model_ModelTy
-  type self   
-  type modelTy   
-end
-module CreusotContracts_Logic_Model_Model_Model_Interface
-  type self   
-  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = self
-  function model (self : self) : ModelTy0.modelTy
-end
-module CreusotContracts_Logic_Model_Model_Model
-  type self   
-  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = self
-  function model (self : self) : ModelTy0.modelTy
-end
-module C02Gnome_Impl0_ModelTy
-  type t   
-  type modelTy  = 
-    t
-end
-module C02Gnome_Impl0_Model_Interface
-  type t   
-  use Type
-  function model (self : Type.c02gnome_ghost t) : t
-end
-module C02Gnome_Impl0_Model
-  type t   
-  use Type
-  function model (self : Type.c02gnome_ghost t) : t
-end
-module C02Gnome_Impl0
-  type t   
-  use Type
-  clone C02Gnome_Impl0_Model as Model0 with type t = t
-  clone C02Gnome_Impl0_ModelTy as ModelTy0 with type t = t
-  clone CreusotContracts_Logic_Model_Model_Model as Model1 with type self = Type.c02gnome_ghost t,
-  type ModelTy0.modelTy = ModelTy0.modelTy, function model = Model0.model
-  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy1 with type self = Type.c02gnome_ghost t,
-  type modelTy = ModelTy0.modelTy
-end
-module C02Gnome_Impl1_Record_Interface
-  type t   
-  use prelude.Prelude
-  use Type
-  clone C02Gnome_Impl0_Model_Interface as Model0 with type t = t
-  val record (a : t) : Type.c02gnome_ghost t
-    ensures { Model0.model result = a }
-    
-end
-module C02Gnome_Impl1_Record
-  type t   
-  use prelude.Prelude
-  use Type
-  clone C02Gnome_Impl0_Model as Model0 with type t = t
-  val record (a : t) : Type.c02gnome_ghost t
-    ensures { Model0.model result = a }
+  type creusotcontracts_logic_ghost_ghost 't = 
+    | CreusotContracts_Logic_Ghost_Ghost opaque_ptr
     
 end
 module CreusotContracts_Logic_Ord_OrdLogic_CmpLog_Interface
@@ -376,6 +319,20 @@ module C02Gnome_Sorted
   predicate sorted (s : Seq.seq t) = 
     SortedRange0.sorted_range s 0 (Seq.length s)
 end
+module CreusotContracts_Logic_Model_Model_ModelTy
+  type self   
+  type modelTy   
+end
+module CreusotContracts_Logic_Model_Model_Model_Interface
+  type self   
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = self
+  function model (self : self) : ModelTy0.modelTy
+end
+module CreusotContracts_Logic_Model_Model_Model
+  type self   
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = self
+  function model (self : self) : ModelTy0.modelTy
+end
 module CreusotContracts_Std1_Vec_Impl0_ModelTy
   type t   
   use seq.Seq
@@ -452,6 +409,31 @@ module CreusotContracts_Logic_Model_Impl1
   clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy1 with type self = borrowed t,
   type modelTy = ModelTy0.modelTy
 end
+module CreusotContracts_Logic_Ghost_Impl0_ModelTy
+  type t   
+  type modelTy  = 
+    t
+end
+module CreusotContracts_Logic_Ghost_Impl0_Model_Interface
+  type t   
+  use Type
+  function model (self : Type.creusotcontracts_logic_ghost_ghost t) : t
+end
+module CreusotContracts_Logic_Ghost_Impl0_Model
+  type t   
+  use Type
+  function model (self : Type.creusotcontracts_logic_ghost_ghost t) : t
+end
+module CreusotContracts_Logic_Ghost_Impl0
+  type t   
+  use Type
+  clone CreusotContracts_Logic_Ghost_Impl0_Model as Model0 with type t = t
+  clone CreusotContracts_Logic_Ghost_Impl0_ModelTy as ModelTy0 with type t = t
+  clone CreusotContracts_Logic_Model_Model_Model as Model1 with type self = Type.creusotcontracts_logic_ghost_ghost t,
+  type ModelTy0.modelTy = ModelTy0.modelTy, function model = Model0.model
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy1 with type self = Type.creusotcontracts_logic_ghost_ghost t,
+  type modelTy = ModelTy0.modelTy
+end
 module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
@@ -459,6 +441,24 @@ end
 module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
+end
+module CreusotContracts_Logic_Ghost_Impl1_Record_Interface
+  type t   
+  use prelude.Prelude
+  use Type
+  clone CreusotContracts_Logic_Ghost_Impl0_Model_Interface as Model0 with type t = t
+  val record (a : t) : Type.creusotcontracts_logic_ghost_ghost t
+    ensures { Model0.model result = a }
+    
+end
+module CreusotContracts_Logic_Ghost_Impl1_Record
+  type t   
+  use prelude.Prelude
+  use Type
+  clone CreusotContracts_Logic_Ghost_Impl0_Model as Model0 with type t = t
+  val record (a : t) : Type.creusotcontracts_logic_ghost_ghost t
+    ensures { Model0.model result = a }
+    
 end
 module CreusotContracts_Logic_Model_Impl0_ModelTy
   type t   
@@ -801,7 +801,7 @@ module C02Gnome_GnomeSort
   clone C02Gnome_SortedRange as SortedRange0 with type t = t, function LeLog0.le_log = LeLog0.le_log
   clone C02Gnome_Sorted as Sorted0 with type t = t, predicate SortedRange0.sorted_range = SortedRange0.sorted_range
   clone CreusotContracts_Logic_Seq_Impl0_PermutationOf as PermutationOf0 with type t = t
-  clone C02Gnome_Impl0_Model as Model1 with type t = borrowed (Type.creusotcontracts_std1_vec_vec t)
+  clone CreusotContracts_Logic_Ghost_Impl0_Model as Model1 with type t = borrowed (Type.creusotcontracts_std1_vec_vec t)
   clone CreusotContracts_Std1_Vec_Impl0_Model as Model2 with type t = t
   clone CreusotContracts_Std1_Vec_Impl0_ModelTy as ModelTy0 with type t = t
   clone CreusotContracts_Logic_Model_Impl1_Model as Model0 with type t = Type.creusotcontracts_std1_vec_vec t,
@@ -811,10 +811,10 @@ module C02Gnome_GnomeSort
   clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve4 with type self = t
   clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve3 with type self = ()
   clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve2 with type self = usize
-  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = Type.c02gnome_ghost (borrowed (Type.creusotcontracts_std1_vec_vec t))
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = Type.creusotcontracts_logic_ghost_ghost (borrowed (Type.creusotcontracts_std1_vec_vec t))
   clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = borrowed (Type.creusotcontracts_std1_vec_vec t)
   clone CreusotContracts_Std1_Ord_Ord_Le_Interface as Le0 with type self = t, function LeLog0.le_log = LeLog0.le_log
-  clone C02Gnome_Impl1_Record_Interface as Record0 with type t = borrowed (Type.creusotcontracts_std1_vec_vec t),
+  clone CreusotContracts_Logic_Ghost_Impl1_Record_Interface as Record0 with type t = borrowed (Type.creusotcontracts_std1_vec_vec t),
   function Model0.model = Model1.model
   clone CreusotContracts_Logic_Model_Impl0_Model as Model3 with type t = Type.creusotcontracts_std1_vec_vec t,
   type ModelTy0.modelTy = ModelTy0.modelTy, function Model0.model = Model2.model
@@ -829,7 +829,7 @@ module C02Gnome_GnomeSort
    = 
   var _0 : ();
   var v_1 : borrowed (Type.creusotcontracts_std1_vec_vec t);
-  var old_v_2 : Type.c02gnome_ghost (borrowed (Type.creusotcontracts_std1_vec_vec t));
+  var old_v_2 : Type.creusotcontracts_logic_ghost_ghost (borrowed (Type.creusotcontracts_std1_vec_vec t));
   var _3 : borrowed (Type.creusotcontracts_std1_vec_vec t);
   var _4 : borrowed (Type.creusotcontracts_std1_vec_vec t);
   var i_5 : usize;
