+++ creusot/tests/should_succeed/vector/03_knuth_shuffle.stdout
@@ -9,9 +9,6 @@ module Type
   use floating_point.Single
   use floating_point.Double
   use prelude.Prelude
-  type c03knuthshuffle_ghost 't = 
-    | C03KnuthShuffle_Ghost opaque_ptr
-    
   type core_marker_phantomdata 't = 
     | Core_Marker_PhantomData
     
@@ -30,62 +27,8 @@ module Type
   type creusotcontracts_std1_vec_vec 't = 
     | CreusotContracts_Std1_Vec_Vec (alloc_vec_vec 't (alloc_alloc_global))
     
-end
-module CreusotContracts_Logic_Model_Model_ModelTy
-  type self   
-  type modelTy   
-end
-module CreusotContracts_Logic_Model_Model_Model_Interface
-  type self   
-  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = self
-  function model (self : self) : ModelTy0.modelTy
-end
-module CreusotContracts_Logic_Model_Model_Model
-  type self   
-  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = self
-  function model (self : self) : ModelTy0.modelTy
-end
-module C03KnuthShuffle_Impl0_ModelTy
-  type t   
-  type modelTy  = 
-    t
-end
-module C03KnuthShuffle_Impl0_Model_Interface
-  type t   
-  use Type
-  function model (self : Type.c03knuthshuffle_ghost t) : t
-end
-module C03KnuthShuffle_Impl0_Model
-  type t   
-  use Type
-  function model (self : Type.c03knuthshuffle_ghost t) : t
-end
-module C03KnuthShuffle_Impl0
-  type t   
-  use Type
-  clone C03KnuthShuffle_Impl0_Model as Model0 with type t = t
-  clone C03KnuthShuffle_Impl0_ModelTy as ModelTy0 with type t = t
-  clone CreusotContracts_Logic_Model_Model_Model as Model1 with type self = Type.c03knuthshuffle_ghost t,
-  type ModelTy0.modelTy = ModelTy0.modelTy, function model = Model0.model
-  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy1 with type self = Type.c03knuthshuffle_ghost t,
-  type modelTy = ModelTy0.modelTy
-end
-module C03KnuthShuffle_Impl1_Record_Interface
-  type t   
-  use prelude.Prelude
-  use Type
-  clone C03KnuthShuffle_Impl0_Model_Interface as Model0 with type t = t
-  val record (a : t) : Type.c03knuthshuffle_ghost t
-    ensures { Model0.model result = a }
-    
-end
-module C03KnuthShuffle_Impl1_Record
-  type t   
-  use prelude.Prelude
-  use Type
-  clone C03KnuthShuffle_Impl0_Model as Model0 with type t = t
-  val record (a : t) : Type.c03knuthshuffle_ghost t
-    ensures { Model0.model result = a }
+  type creusotcontracts_logic_ghost_ghost 't = 
+    | CreusotContracts_Logic_Ghost_Ghost opaque_ptr
     
 end
 module C03KnuthShuffle_RandInRange_Interface
@@ -103,6 +46,20 @@ module C03KnuthShuffle_RandInRange
   val rand_in_range (l : usize) (u : usize) : usize
     ensures { UInt64.to_int l <= UInt64.to_int result && UInt64.to_int result < UInt64.to_int u }
     
+end
+module CreusotContracts_Logic_Model_Model_ModelTy
+  type self   
+  type modelTy   
+end
+module CreusotContracts_Logic_Model_Model_Model_Interface
+  type self   
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = self
+  function model (self : self) : ModelTy0.modelTy
+end
+module CreusotContracts_Logic_Model_Model_Model
+  type self   
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = self
+  function model (self : self) : ModelTy0.modelTy
 end
 module CreusotContracts_Std1_Vec_Impl0_ModelTy
   type t   
@@ -180,6 +137,31 @@ module CreusotContracts_Logic_Seq_Impl0_PermutationOf
   predicate permutation_of (self : Seq.seq t) (o : Seq.seq t) = 
     Permut.permut self o 0 (Seq.length self)
 end
+module CreusotContracts_Logic_Ghost_Impl0_ModelTy
+  type t   
+  type modelTy  = 
+    t
+end
+module CreusotContracts_Logic_Ghost_Impl0_Model_Interface
+  type t   
+  use Type
+  function model (self : Type.creusotcontracts_logic_ghost_ghost t) : t
+end
+module CreusotContracts_Logic_Ghost_Impl0_Model
+  type t   
+  use Type
+  function model (self : Type.creusotcontracts_logic_ghost_ghost t) : t
+end
+module CreusotContracts_Logic_Ghost_Impl0
+  type t   
+  use Type
+  clone CreusotContracts_Logic_Ghost_Impl0_Model as Model0 with type t = t
+  clone CreusotContracts_Logic_Ghost_Impl0_ModelTy as ModelTy0 with type t = t
+  clone CreusotContracts_Logic_Model_Model_Model as Model1 with type self = Type.creusotcontracts_logic_ghost_ghost t,
+  type ModelTy0.modelTy = ModelTy0.modelTy, function model = Model0.model
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy1 with type self = Type.creusotcontracts_logic_ghost_ghost t,
+  type modelTy = ModelTy0.modelTy
+end
 module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
@@ -187,6 +169,24 @@ end
 module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
+end
+module CreusotContracts_Logic_Ghost_Impl1_Record_Interface
+  type t   
+  use prelude.Prelude
+  use Type
+  clone CreusotContracts_Logic_Ghost_Impl0_Model_Interface as Model0 with type t = t
+  val record (a : t) : Type.creusotcontracts_logic_ghost_ghost t
+    ensures { Model0.model result = a }
+    
+end
+module CreusotContracts_Logic_Ghost_Impl1_Record
+  type t   
+  use prelude.Prelude
+  use Type
+  clone CreusotContracts_Logic_Ghost_Impl0_Model as Model0 with type t = t
+  val record (a : t) : Type.creusotcontracts_logic_ghost_ghost t
+    ensures { Model0.model result = a }
+    
 end
 module CreusotContracts_Logic_Model_Impl0_ModelTy
   type t   
@@ -322,7 +322,7 @@ module C03KnuthShuffle_KnuthShuffle
   use prelude.Prelude
   use Type
   clone CreusotContracts_Logic_Seq_Impl0_PermutationOf as PermutationOf0 with type t = t
-  clone C03KnuthShuffle_Impl0_Model as Model1 with type t = borrowed (Type.creusotcontracts_std1_vec_vec t)
+  clone CreusotContracts_Logic_Ghost_Impl0_Model as Model1 with type t = borrowed (Type.creusotcontracts_std1_vec_vec t)
   clone CreusotContracts_Std1_Vec_Impl0_Model as Model2 with type t = t
   clone CreusotContracts_Std1_Vec_Impl0_ModelTy as ModelTy0 with type t = t
   clone CreusotContracts_Logic_Model_Impl1_Model as Model0 with type t = Type.creusotcontracts_std1_vec_vec t,
@@ -333,9 +333,9 @@ module C03KnuthShuffle_KnuthShuffle
   clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve3 with type self = ()
   clone C03KnuthShuffle_RandInRange_Interface as RandInRange0
   clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve2 with type self = usize
-  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = Type.c03knuthshuffle_ghost (borrowed (Type.creusotcontracts_std1_vec_vec t))
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = Type.creusotcontracts_logic_ghost_ghost (borrowed (Type.creusotcontracts_std1_vec_vec t))
   clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = borrowed (Type.creusotcontracts_std1_vec_vec t)
-  clone C03KnuthShuffle_Impl1_Record_Interface as Record0 with type t = borrowed (Type.creusotcontracts_std1_vec_vec t),
+  clone CreusotContracts_Logic_Ghost_Impl1_Record_Interface as Record0 with type t = borrowed (Type.creusotcontracts_std1_vec_vec t),
   function Model0.model = Model1.model
   clone CreusotContracts_Logic_Model_Impl0_Model as Model3 with type t = Type.creusotcontracts_std1_vec_vec t,
   type ModelTy0.modelTy = ModelTy0.modelTy, function Model0.model = Model2.model
@@ -348,7 +348,7 @@ module C03KnuthShuffle_KnuthShuffle
    = 
   var _0 : ();
   var v_1 : borrowed (Type.creusotcontracts_std1_vec_vec t);
-  var old_v_2 : Type.c03knuthshuffle_ghost (borrowed (Type.creusotcontracts_std1_vec_vec t));
+  var old_v_2 : Type.creusotcontracts_logic_ghost_ghost (borrowed (Type.creusotcontracts_std1_vec_vec t));
   var _3 : borrowed (Type.creusotcontracts_std1_vec_vec t);
   var _4 : borrowed (Type.creusotcontracts_std1_vec_vec t);
   var n_5 : usize;
